Move auth.js
