+++ client/api/auth.js
